refactor(svar2)!: delete the pipeline-backed SVAR2 view path
Both routings now slice. Removes Svar2Source, SourceSpec::Svar2 and
run_view_pipeline, along with the temporary differential test that proved
the replacement matches it. The eager Vec<RawRecord> materialization (31 GB
peak for a 0.2 GB output) goes with them. SVAR1's view path is untouched.
